Struct tantivy_common::BitSet

pub struct BitSet { /* private fields */ }

Implementations

impl BitSet

pub fn serialize<T: Write>(&self, writer: &mut T) -> Result<()>

serialize a BitSet.

pub fn with_max_value(max_value: u32) -> BitSet

Create a new BitSet that may contain elements within [0, max_val).

pub fn with_max_value_and_full(max_value: u32) -> BitSet

Create a new BitSet that may contain elements. Initially all values will be set. within [0, max_val).

pub fn clear(&mut self)

Removes all elements from the BitSet.

pub fn intersect_update(&mut self, other: &ReadOnlyBitSet)

Intersect with serialized bitset

pub fn len(&self) -> usize

Returns the number of elements in the BitSet.

pub fn insert(&mut self, el: u32)

Inserts an element in the BitSet

pub fn remove(&mut self, el: u32)

Inserts an element in the BitSet

pub fn contains(&self, el: u32) -> bool

Returns true iff the elements is in the BitSet.

pub fn first_non_empty_bucket(&self, bucket: u32) -> Option<u32>

Returns the first non-empty TinySet associated with a bucket lower or greater than bucket.

Reminder: the tiny set with the bucket bucket, represents the elements from bucket * 64 to (bucket+1) * 64.

pub fn max_value(&self) -> u32

pub fn tinyset(&self, bucket: u32) -> TinySet

Returns the tiny bitset representing the the set restricted to the number range from bucket * 64 to (bucket + 1) * 64.