Struct tantivy_common::TinySet

pub struct TinySet(_);

Implementations

impl TinySet

pub fn serialize<T: Write>(&self, writer: &mut T) -> Result<()>

pub fn into_bytes(self) -> [u8; 8]

pub fn deserialize(data: [u8; 8]) -> Self

pub fn empty() -> TinySet

Returns an empty TinySet.

pub fn full() -> TinySet

Returns a full TinySet.

pub fn clear(&mut self)

pub fn contains(self, el: u32) -> bool

Returns true iff the TinySet contains the element el.

pub fn len(self) -> u32

Returns the number of elements in the TinySet.

pub fn intersect(self, other: TinySet) -> TinySet

Returns the intersection of self and other

pub fn singleton(el: u32) -> TinySet

Creates a new TinySet containing only one element within [0; 64[

pub fn insert(self, el: u32) -> TinySet

Insert a new element within [0..64)

pub fn remove(self, el: u32) -> TinySet

Removes an element within [0..64)

pub fn insert_mut(&mut self, el: u32) -> bool

Insert a new element within [0..64)

returns true if the set changed

pub fn remove_mut(&mut self, el: u32) -> bool

Remove a element within [0..64)

returns true if the set changed

pub fn union(self, other: TinySet) -> TinySet

Returns the union of two tinysets

pub fn is_empty(self) -> bool

Returns true iff the TinySet is empty.

pub fn pop_lowest(&mut self) -> Option<u32>

Returns the lowest element in the TinySet and removes it.

pub fn range_lower(upper_bound: u32) -> TinySet

Returns a TinySet than contains all values up to limit excluded.

The limit is assumed to be strictly lower than 64.

pub fn range_greater_or_equal(from_included: u32) -> TinySet

Returns a TinySet that contains all values greater or equal to the given limit, included. (and up to 63)

The limit is assumed to be strictly lower than 64.